Turn mgc cli version into a variable/input of the run tests workflow
Updating the mgc version today requires changes in two places of a
hardcoded string. And we dont have a way to use the manually triggered
tests with different versions of the mgc.

This patch turns the version into a variable that can be passed as an
argument to the workflow, what can make it easier to write test matrices
including different versions of the cli in the future, and make it
simple to change the default version to be used when new releases comes.