split PR pipeline in two workflows
Tests that takes long time to finish may not be interesting
to block PRs, but we still want to run their categories
regularly. This patch enables 2 more test categories that
were not be ran previously due to contain lots of 
permutations, but make them run just after the test_success
step is finished so we can accept PRs earlier if needed.