feat: add testcases to prosigned url tests
This patch includes testcases of object keys containing slashes
like if the file were in a subfolder path. There was a know issue
related with slashes in the presigned urls that is already fixed
so this test is here to prevent any regression.

This patch modifies an existing "bucket_with_one_object" fixture
to accept an optional indirect parameter to set the object name.
This fixture change should be backwards compatible.